+++ tests/test_models/test_catboost.py
@@ -1,10 +1,12 @@
 import numpy as np
 import pandas as pd
 import pytest
+from catboost import CatBoostRegressor
 
 from etna.datasets.tsdataset import TSDataset
 from etna.models import CatBoostModelMultiSegment
 from etna.models import CatBoostModelPerSegment
+from etna.pipeline import Pipeline
 from etna.transforms.math import LagTransform
 
 
@@ -81,3 +83,24 @@ def test_catboost_multi_segment_forecast(constant_ts):
 
     for segment in forecast.segments:
         assert np.allclose(test[:, segment, "target"], forecast[:, segment, "target"])
+
+
+def test_get_model_multi():
+    etna_model = CatBoostModelMultiSegment()
+    model = etna_model.get_model()
+    assert isinstance(model, CatBoostRegressor)
+
+
+def test_get_model_per_segment_before_training():
+    etna_model = CatBoostModelPerSegment()
+    with pytest.raises(ValueError, match="Can not get the dict with base models, the model is not fitted!"):
+        _ = etna_model.get_model()
+
+
+def test_get_model_per_segment_after_training(example_tsds):
+    pipeline = Pipeline(model=CatBoostModelPerSegment(), transforms=[LagTransform(in_column="target", lags=[2, 3])])
+    pipeline.fit(ts=example_tsds)
+    models_dict = pipeline.model.get_model()
+    assert isinstance(models_dict, dict)
+    for segment in example_tsds.segments:
+        assert isinstance(models_dict[segment], CatBoostRegressor)