+++ tests/test_models/test_simple_models.py
@@ -5,6 +5,8 @@ import pytest
 from etna.models.moving_average import MovingAverageModel
 from etna.models.naive import NaiveModel
 from etna.models.seasonal_ma import SeasonalMovingAverageModel
+from etna.models.seasonal_ma import _SeasonalMovingAverageModel
+from etna.pipeline import Pipeline
 
 
 @pytest.mark.parametrize("model", [SeasonalMovingAverageModel, NaiveModel, MovingAverageModel])
@@ -93,3 +95,36 @@ def test_moving_average_forecaster_correct(simple_df):
     answer = answer.sort_values(by=["segment", "timestamp"])
 
     assert np.all(res.values == answer.values)
+
+
+@pytest.mark.parametrize(
+    "etna_model_class",
+    (
+        SeasonalMovingAverageModel,
+        MovingAverageModel,
+        NaiveModel,
+    ),
+)
+def test_get_model_before_training(etna_model_class):
+    """Check that get_model method throws an error if per-segment model is not fitted yet."""
+    etna_model = etna_model_class()
+    with pytest.raises(ValueError, match="Can not get the dict with base models, the model is not fitted!"):
+        _ = etna_model.get_model()
+
+
+@pytest.mark.parametrize(
+    "etna_model_class,expected_class",
+    (
+        (NaiveModel, _SeasonalMovingAverageModel),
+        (SeasonalMovingAverageModel, _SeasonalMovingAverageModel),
+        (MovingAverageModel, _SeasonalMovingAverageModel),
+    ),
+)
+def test_get_model_after_training(example_tsds, etna_model_class, expected_class):
+    """Check that get_model method returns dict of objects of _SeasonalMovingAverageModel class."""
+    pipeline = Pipeline(model=etna_model_class())
+    pipeline.fit(ts=example_tsds)
+    models_dict = pipeline.model.get_model()
+    assert isinstance(models_dict, dict)
+    for segment in example_tsds.segments:
+        assert isinstance(models_dict[segment], expected_class)